requirements.txt
speedtest-cli
requests
requests-toolbelt
